refactor: refactor execution function
