+++ .eslintrc
@@ -63,7 +63,7 @@
         "prefer-const": [
             "error",
             {
-                "destructuring": "any",
+                "destructuring": "all",
                 "ignoreReadBeforeAssign": false
             }
         ],
