+++ src/utils/execution.js
@@ -7,40 +7,33 @@ const execution = (previousOperator, stacks) => {
     const previousOperand = +numberStack[numberStack.length - 2];
     const lastOperand = +numberStack[numberStack.length - 1];
 
+    operatorsStack = removeItemFromOperatorStack(operatorsStack, 1);
+    if (previousOperator && previousOperator.trim().match(/[√^]/g)) {
+        numberStack = removeItemFromNumberStack(numberStack, 1);
+    } else if (previousOperator) {
+        numberStack = removeItemFromNumberStack(numberStack, 2);
+    }
+
     switch (previousOperator && previousOperator.trim()) {
         case '^':
-            numberStack = removeItemFromNumberStack(numberStack, 1);
-            operatorsStack = removeItemFromOperatorStack(operatorsStack, 1);
             numberStack.push(pow(lastOperand).toString());
             break;
         case '√':
-            numberStack = removeItemFromNumberStack(numberStack, 1);
-            operatorsStack = removeItemFromOperatorStack(operatorsStack, 1);
             numberStack.push(sqrt(lastOperand).toString());
             break;
         case '/':
-            numberStack = removeItemFromNumberStack(numberStack, 2);
-            operatorsStack = removeItemFromOperatorStack(operatorsStack, 1);
             numberStack.push(division(previousOperand, lastOperand).toString());
             break;
         case '*':
-            numberStack = removeItemFromNumberStack(numberStack, 2);
-            operatorsStack = removeItemFromOperatorStack(operatorsStack, 1);
             numberStack.push(multiplication(previousOperand, lastOperand).toString());
             break;
         case '+':
-            numberStack = removeItemFromNumberStack(numberStack, 2);
-            operatorsStack = removeItemFromOperatorStack(operatorsStack, 1);
             numberStack.push(addition(previousOperand, lastOperand).toString());
             break;
         case '-':
-            numberStack = removeItemFromNumberStack(numberStack, 2);
-            operatorsStack = removeItemFromOperatorStack(operatorsStack, 1);
             numberStack.push(subtraction(previousOperand, lastOperand).toString());
             break;
         case '%':
-            numberStack = removeItemFromNumberStack(numberStack, 2);
-            operatorsStack = removeItemFromOperatorStack(operatorsStack, 1);
             numberStack.push(modulo(previousOperand, lastOperand).toString());
             break;
