+++ src/utils/warningMessage.js
@@ -1,7 +1,7 @@
 import { bindActionCreators } from 'redux';
 
 import { WARNING_MESSAGE_TIMEOUT } from '@/constants';
-import { setOwnValue } from '@/redux/actions/actions';
+import { setOwnValue } from '@/redux/actions';
 import { store } from '@/redux/store';
 
 const { dispatch } = store;
